pyproject.toml
[build-system]
requires = ["setuptools>=69.5.1", "wheel>=0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "workflows"
version = "0.1.0"
description = "Reusable GitHub Actions workflows and automation infrastructure"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Workflow Contributors"}
]
keywords = ["github-actions", "workflows", "ci-cd", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Repository = "https://github.com/stranske/Workflows"
Issues = "https://github.com/stranske/Workflows/issues"

[project.optional-dependencies]
dev = [
    "pre-commit==4.4.0",
    "black==25.11.0",
    "ruff==0.14.7",
    "isort==7.0.0",
    "docformatter==1.7.7",
    "mypy==1.19.0",
    "flake8==7.3.0",
    "pytest==9.0.1",
    "pytest-cov==7.0.0",
    "coverage==7.12.0",
    "pytest-xdist==3.8.0",

    # Required by repo dependency checks / CI fallback checks
    "hypothesis>=6.0.0",
    "pandas",
    "numpy",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "jsonschema>=4.0.0",
    "PyYAML>=6.0.0",

    # Used by scripts/sync_test_dependencies.py to edit pyproject.toml
    "tomlkit>=0.13.0",
]

[tool.setuptools]
# This repo is primarily automation + scripts; avoid setuptools trying to auto-discover
# random top-level dirs as importable packages (which breaks editable installs).
packages = []

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # Exclude standard directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | archive
  | \.extraction
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = []
skip_glob = ["archive/*", ".extraction/*"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = []

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # Line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disable_error_code = "import-untyped"
exclude = [
    "^archive/",
    "^\\.extraction/",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "integration: marks tests as integration tests (may require external resources)",
    "cosmetic: marks tests for cosmetic/formatting changes",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial: marks tests that must run serially",
    "xdist_group: marks tests that must share a single xdist worker",
]
filterwarnings = [
    "error",
]

[tool.coverage.run]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/.venv/*",
    "archive/*",
    ".extraction/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
